requirements.txt
flask==3.0.2
python-telegram-bot[http2]==20.8
python-dotenv==1.0.1
requests==2.31.0
werkzeug==3.0.1
flask_wtf==1.2.1
httpx==0.26.0
aiohttp==3.9.3
aiofiles==23.2.1
secure==0.3.0
Pillow==10.2.0
beautifulsoup4==4.12.2
Jinja2==3.1.3
itsdangerous==2.1.2
click==8.1.7 